implementacion de nueva interfaz de sizes
