+++ app/src/main/java/org/compose_projects/socialocal/ui/components/textField/config/TextFieldSizes.kt
@@ -1,0 +1,13 @@
+package org.compose_projects.socialocal.ui.components.textField.config
+
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.ui.unit.Dp
+
+interface TextFieldSizes {
+    val height: Dp
+    val width: Dp
+    val shape: RoundedCornerShape
+    val maxLines: Int
+    val maxChar: Int
+
+}
